test: extract voteForOpenQuestion
